requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
cloudinary==1.28.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.9
django-bootstrap-v5==1.0.7
django-bootstrap5==21.3
django-crispy-forms==1.14.0
django-emoji-picker==0.0.6
django-heroku==0.3.1
django-registration==3.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.10.1
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
